Minor refactoring changes
